feat(promo-codes): add promo codes to orders
- added feign clients of Promo Code, Promo Stat APIs
- added feign client of Product API to fetch and manage order item price

+++ order-service/src/main/java/com/mealmap/orderservice/document/value/OrderItem.java
@@ -10,7 +10,7 @@ import lombok.Data;
 public class OrderItem {
     private Long productId;
 
-    private Long priceWhenOrdered;
+    private Integer priceWhenOrdered;
 
     private Integer quantity;
 }
